==> generator/1_0_orders_and_products.ipynb <==
# 先从产品以及相关的订单销售数据开始

from imports import *
from utils import *

order_m_path = os.path.join(BASE_DIR, 'order_monthly_template.xlsx')
order_m = pd.read_excel(order_m_path)

order_m

order_m['year'] = order_m['month'].apply(lambda x: x.year)

# 生成数据

# from yearly data to monthly breakdown

# grow = 100
# total_year = intercept * 12 + grow * 12/2
# total_year

# 读取年度数据
df = pd.read_excel(BASE_INFO_PATH, sheet_name= '产品及销量')

df

df_price = df.loc[5:8]

df_price


df_price_adj = df_price.melt(id_vars=['产品'], var_name='year', value_name='count').dropna()
df_price_adj = pd.merge(df_price_adj,get_product()[['product_id','product_code']], left_on = '产品', right_on='product_code', how='left')
df_price_adj

'product_id == {} and year = {}'.format(1, 2018)

# p = df_price.iloc[0,:].values[1:].reshape(-1,1)
# _id = np.array([i for i in range(1, len(p)+1)]).reshape(-1,1)
# intercept, slope = linear_regression_one_variable(_id,p)

# 

def gen_rand(l, total_year,intercept = None, growth = 'linear'):
    """生成随机月度销量数据"""
    
    # total_year = 20000


    if intercept is None:
        intercept = int(total_year * 0.05)

    # exponential growth
    if growth == 'exponential':
        yearly_growth = 0.4
        growth_rate_monthly = np.power(1 + yearly_growth, 1/12) - 1
    # linear growth
    if growth == 'linear':
        yearly_base = intercept * 12
        yearly_rest= total_year - yearly_base
        year_end_growth = yearly_rest * 2 / 12
        growth_count = year_end_growth / 12
        

    
    var = 0.2

    
    g = []
    
    # TODO P3: redo this section
    for i in range(l):
        if i == 0:
            base_line = intercept
        else:
            # base_line = intercept + growth_rate_monthly
            if growth == 'exponential':
                base_line = base_line * (1 + growth_rate_monthly)
            if growth == 'linear':
                base_line = intercept + growth_count * i
            
        if growth == 'exponential':
            _ = base_line * (1+np.random.normal(loc = 0, scale = var))
        
        if growth == 'linear':
            _ = base_line * (1+np.random.normal(loc = 0, scale = var))
            
        _ = np.round(_)
        _ = int(_)
        
        # check non-negative
        if _<0:
            _ = 0
            
        g.append(_)
        
    return g

df_g_list = []


for product in order_m['product_id'].unique():
    # product
    df_this_prod = order_m[order_m['product_id'] == product]
    
    last_month_of_this_product = 0
    
    for idx, year in enumerate(df_this_prod['year'].unique()):


        year_total = df_price_adj.query('product_id == {} and year == {}'.format(product,year))['count'].values[0]
        
        
        print(product,year, year_total)
        
        df_this_year = df_this_prod[df_this_prod['year'] == year]
        
        l = df_this_year.shape[0]
        
        if idx > 0:
            # intercept use last year
            intercept = last_month_of_this_product
        else:
            # none means 5% of 1st year
            intercept = None
        
        df_this_year['count_raw'] = gen_rand(l, year_total, intercept = intercept)
        last_month_of_this_product = df_this_year['count_raw'].values[-1]
        
        df_g_list.append(df_this_year)

df_order_g = pd.concat(df_g_list)

df_order_g

px.line(df_order_g, x = 'month', y = 'count_raw', color = 'product_id')

df_order_g

# group by product_id and month
df_order_g['year'] = df_order_g['month'].apply(lambda x: x.year)

df_order_g

# 年度以及月度产量

df_order_g.groupby(['year', 'product_id']).sum()

# 保存

df_order_g.to_csv(os.path.join(GE_DIR, 'order_monthly.csv'), index = False)
