==> generator/1_3_orders_detail_region.ipynb <==
# 把 order 进行细节的拆分 拆到每一个order的细节 具备





1. order_id
1. order_date
1. order_datetime, 具体的日期，用于分析行为
2. order_user_id，留在后边用
3. order_provice
4. order_city
    这两个部分先按照省份来拆分
5. 


# 2023/02/07 只生成了部分 12月份的数据
# 2023/5 改成生成全部的细节数据 还未完全完成

from imports import *

# read order monthly data 

df = pd.read_csv(os.path.join(GE_DIR,'order_monthly.csv'))

df

week_ratio = [0.8,1,2,2.5,3.8,4,3.5]

# import relativedelta
from dateutil.relativedelta import relativedelta

# Generate random hour and minute according to the ratio of 24 hours

random_24_hour_rato = [1,1,1,1,1,1,2,3,4,8,7,6,3,6,5,4,3,3,2,4,7,8,12,7]
px.scatter(random_24_hour_rato)

# 按照平均，周末更多的方式分配订单


df_list = []
# iterate over df
for idx, row in df.iterrows():
    
    month_str = row['month']
    count_raw = int(row['count_raw'])
    
    first_day = datetime.strptime(month_str, '%Y-%m-%d')
    this_month = first_day.month
    last_day = first_day + relativedelta(months=1, days=-1)
    
    df_this_month = pd.DataFrame()
    dates_of_this_month = pd.date_range(first_day, last_day)
    df_this_month['date'] = dates_of_this_month

    week_days = dates_of_this_month.dayofweek + 1
    df_this_month['week_day'] = week_days
    
    # Generate order count for each week day according to week_ratio
    order_count = [count_raw * ratio / sum(week_ratio) for ratio in week_ratio]
    
    # Generate dict including day of week and order count
    _dict = dict(zip(week_days, order_count))
    
    # number of week days of each week_day
    df_this_month['num_week_day'] = df_this_month['week_day'].map(df_this_month['week_day'].value_counts())
    
    df_this_month['count_total'] = df_this_month['week_day'].map(_dict)
    
    df_this_month['count_orders'] = df_this_month['count_total'] / df_this_month['num_week_day']
    
    # Generate randomness for count_orders follow normal distribution
    df_this_month['count_orders_this_day'] = df_this_month['count_orders'].apply(lambda x: int(np.random.normal(x, x/5)))
    
    df_list.append(df_this_month)


df_this_month

df_all = pd.concat(df_list)

df_all

# Generate random hour and minute according to the ratio of 24 hours
def generate_hour_minute():
    hour = np.random.choice(range(24), p=np.array(random_24_hour_rato)/sum(random_24_hour_rato))
    minute = np.random.choice(range(60))
    return hour, minute

# Generate id with random hour and minute

# 
order_list = []

idx = 0
# for idx, row in df_this_month.iterrows():

for idx, row in df_all.iterrows():

    for i in range(row['count_orders_this_day']):
        hour, minute = generate_hour_minute()
        
        order = {
            "id": idx,
            "date": row['date'].strftime('%Y-%m-%d'),
            "time": f"{hour:02d}:{minute:02d}",
            "datetime": row['date'].strftime('%Y-%m-%d') + f" {hour:02d}:{minute:02d}",
            
        }
        
        idx += 1
        
        order_list.append(order)

# change to df
df_order = pd.DataFrame(order_list)

df_order

# 获取省份





# Generate list of all provices in China
provices = ['北京市','天津市','河北省','山西省','内蒙古自治区','辽宁省','吉林省','黑龙江省','上海市','江苏省','浙江省','安徽省','福建省','江西省','山东省','河南省','湖北省','湖南省','广东省','广西壮族自治区','海南省','重庆市','四川省','贵州省','云南省','西藏自治区','陕西省','甘肃省','青海省','宁夏回族自治区','新疆维吾尔自治区','台湾省','香港特别行政区','澳门特别行政区']

# generate all GDP data of all provices

# read order_propotion data

df_order_propotion = pd.read_excel(os.path.join(BASE_DIR,'order_propotion.xlsx'))

df_order_propotion

# generate n random choise from list df_order_propotion['provice']
import random

# generate n random choise from list df_order_propotion['provice'] with probabiliy of df_order_propotion['propotion']
import random

# generate n random choices from list df_order_propotion['province'] with probability of df_order_propotion['proportion']
n = 100
provinces = df_order_propotion['provice']
random_choices = random.choices(provinces, weights=df_order_propotion['propotion'], k=n)

# print the random choices
print(random_choices)


# generate random provice base on the propotion

def generate_provice(x):
    # make sum of propotion equal to 1
    df_order_propotion['propotion'] = df_order_propotion['propotion'] / sum(df_order_propotion['propotion'])
    
    provice = np.random.choice(df_order_propotion['provice'], p=df_order_propotion['propotion'].values)
    return provice

def generate_provice_m(n):
    # n = 100
    provinces = df_order_propotion['provice']
    random_choices = random.choices(provinces, weights=df_order_propotion['propotion'], k=n)

    # print the random choices
    # print(random_choices)
    return random_choices

# df_order['provice'] = df_order['id'].apply(generate_provice)

province_generated = generate_provice_m(df_order.shape[0])


df_order['provice'] = province_generated

df_order

#
# df_order.groupby('provice').count()

df_order

# read city data

df_city = pd.read_excel('../map_generation_src/省市县区.xlsx')

# remove "省" and "市" from provice
df_city['provice'] = df_city['provice'].apply(lambda x: x.replace('省','').replace('市',''))

df_city['provice'].unique()

df_city

df_city

# Generate random city of each provice
# too slow
def generate_city(x,df_city = df_city):
    try:
        df_this_provice = df_city[df_city['provice'] == x]
        city = np.random.choice(df_this_provice['city'])
    except:
        city = x
    return city


def generate_city(n,p,df_city = df_city):
    # n = 100
    df_this_provice = df_city[df_city['provice'] == p]
    city = np.random.choice(df_this_provice['city'], k = n)
    

    # print the random choices
    # print(random_choices)
    return random_choices

#

for p in df_order['provice'].unique():
    
    df_p = df_order[df_order['provice'] == p]
    
    # generate city for this p
    city_generated = generate_city(df_p.shape[0])

df_order['city'] = df_order['provice'].apply(generate_city)

df_order

# save as csv
df_order.to_csv(os.path.join(GE_DIR,'order_detail.csv'), index=False)
