==> generator/1_4_order_prediction.ipynb <==
# 输出销量的部分预测

1. 过去的销量
2. 过去的生产、库存
3. （对于未来的预测）共同决定未来的生产

from imports import *

# read order details from GE_DIR

order = pd.read_csv(os.path.join(GE_DIR, 'order_detail.csv'))

order

# groupby date, provice
# df = order.groupby(['date', 'provice']).count().reset_index()
df_by_date = order.groupby('date').count().reset_index()
df_by_date

px.line(df_by_date, x='date', y='id', color='provice')

# Prediciton with ARIMA


# extend dates
def extend_dates(last_date, n = 30):
    
    # last_date = dates[-1]
    # next day
    x = [datetime.strptime(last_date, '%Y-%m-%d') + timedelta(days=i) for i in range(n)]
    x = [datetime.strftime(x,'%Y-%m-%d') for x in x]
        
    return x

# Use ARIMA model to predict order count

from statsmodels.tsa.arima.model import ARIMA

# df_for_pred = df[df['provice'] == '上海'].reset_index(drop=True)
df_for_pred = df_by_date


def pred(df_for_pred):
    # use data from 上海
    data = df_for_pred['id'].to_list()
    train = data[:-10]
    test = data[-10:]
    # train, test
    model = ARIMA(train, order=(2,1,2)).fit()  # ARIMA = (2,1,2)

    delta = model.fittedvalues - train

    # score = 1 - delta.var()/train[1:].var()
    pred = model.predict(start=1, end=len(train)+10)
    
    # confidence 
    conf = model.conf_int(alpha=0.05)
    
    df_data = pd.DataFrame()
    df_data['date'] = df_for_pred['date']
    df_data['value'] = data
    df_data['source'] = 'History'
    df_data

    # extend dates

    dates = df_data['date'].values
    last_date = dates[-1]
    
    df_pred = pd.DataFrame()
    df_pred['date'] = extend_dates(last_date, len(pred))
    df_pred['value'] = pred
    df_pred['source'] = 'Prediction'
    df_all = pd.concat([df_data, df_pred], axis=0)
    
    return df_all, model

df_all, model = pred(df_for_pred)

forecast = model.get_forecast(30)
yhat = forecast.predicted_mean
yhat_conf_int = forecast.conf_int(alpha=0.1)

# change to dataframe
yhat_conf_int_df = pd.DataFrame(yhat_conf_int)

yhat_conf_int_df

yhat_conf_int_df.columns = ['lower', 'upper']

# plot
fig = px.line(df_all, x='date', y='value', color='source')

fig.add_trace(go.Scatter(x=df_all['date'], y=yhat, name='Prediction_mean'))
fig.add_trace(go.Scatter(x=df_all['date'], y=yhat_conf_int_df['lower'], name='Lower'))
fig.add_trace(go.Scatter(x=df_all['date'], y=yhat_conf_int_df['upper'], name='Upper'))

# plot the result

px.line(df_all, x='date', y='value', color='source')

# Save the result as csv to GE_DIR
df_all.to_csv(os.path.join(GE_DIR, 'order_prediction.csv'), index=False)

# Prediction with Prophet  (Not Use for now)

# !pip install prophet
# !pip install fbprophet

from prophet import Prophet

# fit prophet model
model = Prophet()

model.fit()
