==> generator/6_3_oee_cycletime.ipynb <==
# read eqipment data


from imports import *

df = pd.read_csv(os.path.join(BASE_DIR,'equipment.csv'))

df

# read target cycle time


target = 80


# 每条线基本在80上下

df.station.unique()

# 增加以一个配置文件 用于存储每个设备的cycletime 相关的信息

# 随机生成每条线的cycle time
df_cycle = pd.DataFrame()

# for line 1 

df_cycle['station'] = df.station.unique()
df_cycle['cycle_time'] = np.random.randint(60,85,df.station.nunique())
df_cycle['line'] = 1

# for line 2
# 增加到90秒
df_cycle2 = df_cycle.copy()
df_cycle2['line'] = 2
df_cycle2['cycle_time'] = np.random.randint(60,90,df.station.nunique())

# all cycle time
df_cycle = pd.concat([df_cycle,df_cycle2],axis=0)

df_cycle

# save to csv
df_cycle.to_csv(os.path.join(GE_DIR,'cycle_time.csv'),index=False)
