==> generator/7_1_equipment_center_feature.ipynb <==
# 获取数据
# 做一个参数 用于表征设备状态，未使用

from imports import *

# read equipment_status from GE dir

equipment_status = pd.read_csv(os.path.join(GE_DIR,'equipment_vibration.csv'))

equipment_status = equipment_status.reset_index()
equipment_status

rolling_std = equipment_status['noise_data'].rolling(window=1000, step=1000).std()
rolling_std = rolling_std.reset_index()


rolling_std

# rename columns in rolling_std
rolling_std.rename(columns={'noise_data':'rolling_std'}, inplace=True)
rolling_std

rolling_std = rolling_std.reset_index()

# time data
df = pd.merge(rolling_std, equipment_status, how='left', on ='index')
df

px.scatter(df, x='index', y='rolling_std')

# save to csv
df.to_csv(os.path.join(GE_DIR,'equipment_vibration_rolling_std.csv'), index=False)
