==> generator/9_1_supplier_delivery.ipynb <==
# 生成订单以后，供应商的每一次运货的情况，是不是本身能够一直到货，
# 用于评价供应商的运货能力

# 2022/03 目前低优先级


# 从BOM Usage 出发 生成供应商的订单，（这个在上一个notebook已经完成）
# 
# 9_3这个notebook
# 产生供应商的运货情况


from imports import *

# BOM读取

# read bom

bom = pd.read_excel(os.path.join(BASE_DIR,'bom.xlsx'))

bom

# Change columns of A,B,C,D in bom to rows

bomb = pd.melt(bom, id_vars=['名称'], value_vars=['A','B','C','D'], var_name='part_name', value_name='value')

bomb.dropna(inplace=True)

bomb.drop('value', axis=1, inplace=True)

bomb

#
